fix(persistence): improve table state management and prevent duplicate names
- fix: update loadLatestTableState to load all tables instead of just the most recent one
- fix: modify deleteTable to also delete the table from the database
- feat: add validation to prevent creating or renaming tables with duplicate names
- fix: add trailing slashes to API endpoints to match backend routes
- fix: disable automatic redirects in FastAPI to eliminate 307 redirect logs
- refactor: remove console logs for normal operations to keep the console clean
- fix: improve error handling in table creation and renaming modals
- docs: add clear error messages for duplicate table names

+++ backend/src/app/core/auth.py
@@ -35,7 +35,7 @@ class Token(BaseModel):
 
 class TokenData(BaseModel):
     """Token data model."""
-    exp: int
+    exp: float  
 
 
 def verify_password(password: str) -> bool:
